+++ app/screens/CareerCounsellor/CareerCounsellor.js
@@ -10,7 +10,7 @@ import {
 } from 'react-native';
 import { Divider } from 'react-native-elements';
 
-import realm from '../../schema';
+import realm from '../../db/schema';
 import User from '../../utils/user';
 import uuidv4 from '../../utils/uuidv4';
