+++ app/screens/CareerPlanning/game_history/personality_report.js
@@ -11,7 +11,7 @@ import { Divider } from 'react-native-elements';
 import mainStyles from '../../../assets/style_sheets/main/main';
 import StatusBar from '../../../components/shared/status_bar';
 import characteristicList from '../../../data/json/characteristic_jobs';
-import realm from '../../../schema';
+import realm from '../../../db/schema';
 import User from '../../../utils/user';
 
 export default class PersonalityReport extends Component {
