+++ app/screens/CareerPlanning/game_history/recommendation_report.js
@@ -10,7 +10,7 @@ import {
 
 import mainStyles from '../../../assets/style_sheets/main/main';
 
-import realm from '../../../schema';
+import realm from '../../../db/schema';
 import User from '../../../utils/user';
 import subjectList from '../../../data/json/subject';
 import characteristicList from '../../../data/json/characteristic_jobs';
