+++ app/screens/CareerPlanning/personality_jobs_screen.android.js
@@ -18,7 +18,7 @@ import MaterialIcon from 'react-native-vector-icons/MaterialIcons';
 import styles from '../../assets/style_sheets/profile_form';
 import headerStyles from '../../assets/style_sheets/header';
 import shareStyles from './style';
-import realm from '../../schema';
+import realm from '../../db/schema';
 import User from '../../utils/user';
 import characteristicList from '../../data/json/characteristic_jobs';
