+++ app/screens/CareerPlanning/personality_jobs_screen.ios.js
@@ -19,7 +19,7 @@ import MaterialIcon from 'react-native-vector-icons/MaterialIcons';
 import mainStyles from '../../assets/style_sheets/main/main';
 import headerStyles from '../../assets/style_sheets/header';
 import shareStyles from './style';
-import realm from '../../schema';
+import realm from '../../db/schema';
 import User from '../../utils/user';
 import characteristicList from '../../data/json/characteristic_jobs';
