+++ app/screens/CareerPlanning/recommendation_screen.ios.js
@@ -16,7 +16,7 @@ import headerStyles from '../../assets/style_sheets/header';
 import shareStyles from './style';
 import MaterialIcon from 'react-native-vector-icons/MaterialIcons';
 
-import realm from '../../schema';
+import realm from '../../db/schema';
 import User from '../../utils/user';
 import subjectList from '../../data/json/subject';
 import characteristicList from '../../data/json/characteristic_jobs';
