+++ app/screens/CareerPlanning/subject_screen.android.js
@@ -22,7 +22,7 @@ import FooterBar from '../../components/FooterBar';
 import styles from '../../assets/style_sheets/profile_form';
 import shareStyles from './style';
 
-import realm from '../../schema';
+import realm from '../../db/schema';
 import User from '../../utils/user';
 import uuidv4 from '../../utils/uuidv4';
