+++ app/screens/PersonalUnderstanding/PersonalUnderstandingForm.js
@@ -23,7 +23,7 @@ import CloseButton from '../../components/shared/close_button';
 import SaveButton from '../../components/shared/save_button';
 
 import Form from './_Form';
-import realm from '../../schema';
+import realm from '../../db/schema';
 import User from '../../utils/user';
 import uuidv4 from '../../utils/uuidv4';
 import styles from './styles';
